pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tidus"
version = "1.1.0"
description = "Vendor-agnostic, enterprise-grade, cost-aware AI model router and governance system"
readme = "README.md"
license = { text = "Apache-2.0" }
authors = [{ name = "Kenny Wong", email = "lapkei01@gmail.com" }]
keywords = ["ai", "llm", "router", "cost-optimization", "enterprise", "governance"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    # Data validation
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.30",
    "aiosqlite>=0.20.0",
    "alembic>=1.13.0",
    # HTTP client
    "httpx>=0.27.0",
    # Vendor SDKs
    "openai>=1.35.0",
    "anthropic>=0.30.0",
    "google-generativeai>=0.7.0",
    "mistralai>=1.0.0",
    # Token counting
    "tiktoken>=0.7.0",
    "sentencepiece>=0.2.0",
    # Scheduling (health probes + price sync)
    "apscheduler>=3.10.0",
    # MCP
    "mcp>=1.0.0",
    # YAML
    "pyyaml>=6.0.1",
    # Utilities
    "python-dotenv>=1.0.0",
    "structlog>=24.1.0",
    # Auth / SSO
    "pyjwt[crypto]>=2.8.0",
    "prometheus-fastapi-instrumentator>=7.1.0",
    # Rate limiting
    "slowapi>=0.1.9",
    "prometheus-client>=0.24.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "respx>=0.21.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]
load = [
    "locust>=2.29.0",
]

[project.scripts]
tidus = "tidus.main:run"
tidus-mcp = "tidus.mcp.server:main"
tidus-setup = "tidus.cli.setup_wizard:main"

[project.urls]
Homepage = "https://github.com/kensterinvest/tidus"
Repository = "https://github.com/kensterinvest/tidus"
Issues = "https://github.com/kensterinvest/tidus/issues"
Documentation = "https://github.com/kensterinvest/tidus/tree/main/docs"

[tool.hatch.build.targets.wheel]
packages = ["tidus"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = [
    "E501",   # line too long — descriptions and test assertions are intentionally verbose
    "UP042",  # str+Enum → StrEnum — deferred; changing 7 production enums is out of scope
]

[tool.mypy]
python_version = "3.12"
strict = false
ignore_missing_imports = true

[dependency-groups]
dev = [
    "anyio>=4.13.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "respx>=0.22.0",
]
